+++ src/GS/ApiBundle/Controller/TopicController.php
@@ -6,8 +6,10 @@ use Symfony\Component\HttpFoundation\Request;
 use FOS\RestBundle\Controller\FOSRestController;
 use FOS\RestBundle\Controller\Annotations\RouteResource;
 
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
+
 use GS\ApiBundle\Entity\Topic;
-use GS\ApiBundle\Entity\Address;
+use GS\ApiBundle\Entity\Registration;
 
 /**
  * @RouteResource("Topic", pluralize=false)
@@ -15,33 +17,75 @@ use GS\ApiBundle\Entity\Address;
 class TopicController extends FOSRestController
 {
 
-    public function deleteAction($id)
+    /**
+     * @Security("has_role('ROLE_USER')")
+     */
+    public function postAction(Request $request)
     {
-        $em = $this->getDoctrine()->getManager();
-        
-        $topic = $em
-            ->getRepository('GSApiBundle:Topic')
-            ->find($id)
-            ;
+        $form = $this->get('gsapi.form_generator')->getTopicForm(null, 'post_topic');
+        $this->denyAccessUnlessGranted('create', $form->getData());
+        $form->handleRequest($request);
 
-        $em->remove($topic);
-        $em->flush();
+        if ($form->isValid()) {
+            $topic = $form->getData();
+            $topic->addOwner($this->getUser());
+            $activity = $topic->getActivity();
+            $activity->addTopic($topic);
+            
+            $em = $this->getDoctrine()->getManager();
+            $em->persist($topic);
+            $em->flush();
 
-        $view = $this->view(array(), 200);
+            $view = $this->view(array('id' => $topic->getId()), 200);
+            
+        } else {
+            $view = $this->get('gsapi.form_generator')->getFormView($form);
+        }
         return $this->handleView($view);
     }
 
-    public function getAction($id)
+    /**
+     * @Security("is_granted('delete', topic)")
+     */
+    public function removeAction(Topic $topic)
     {
-        $topic = $this->getDoctrine()->getManager()
-            ->getRepository('GSApiBundle:Topic')
-            ->find($id)
-            ;
+        $form = $this->get('gsapi.form_generator')->getTopicDeleteForm($topic);
+        $view = $this->get('gsapi.form_generator')->getFormView($form);
+        return $this->handleView($view);
+    }
 
+    /**
+     * @Security("is_granted('delete', topic)")
+     */
+    public function deleteAction(Topic $topic, Request $request)
+    {
+        $form = $this->get('gsapi.form_generator')->getTopicDeleteForm($topic);
+        $form->handleRequest($request);
+        
+        if ($form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($topic);
+            $em->flush();
+
+            $view = $this->view(null, 204);
+        } else {
+            $view = $this->getFormView($form);
+        }
+        return $this->handleView($view);
+    }
+
+    /**
+     * @Security("is_granted('view', topic)")
+     */
+    public function getAction(Topic $topic)
+    {
         $view = $this->view($topic, 200);
         return $this->handleView($view);
     }
 
+    /**
+     * @Security("has_role('ROLE_USER')")
+     */
     public function cgetAction()
     {
         $listTopics = $this->getDoctrine()->getManager()
@@ -53,106 +97,60 @@ class TopicController extends FOSRestController
         return $this->handleView($view);
     }
 
-    public function postAction(Request $request)
+    /**
+     * @Security("is_granted('edit', topic)")
+     */
+    public function editAction(Topic $topic)
     {
-        if (0 === strpos($request->headers->get('Content-Type'), 'application/json')) {
-            $data = json_decode($request->getContent(), true);
-        }
-        else
-        {
-            $view = $this->view(array(), 301);
-            return $this->handleView($view);
-        }
-        
-        $em = $this->getDoctrine()->getManager();
-
-        $topic = new Topic();
-        if(! $this->setTopicData($em, $topic, $data))
-        {
-            $view = $this->view(array(), 301);
-            return $this->handleView($view);
-        }
-
-        $em->persist($topic);
-        $em->flush();
-
-        if(null != $topic->getId()) {
-            $view = $this->view(array('id' => $topic->getId()), 200);
-        }
-        else
-        {
-            $view = $this->view(array(), 301);
-        }
+        $form = $this->get('gsapi.form_generator')->getTopicForm($topic, 'put_topic', 'PUT');
+        $view = $this->get('gsapi.form_generator')->getFormView($form);
         return $this->handleView($view);
     }
 
-    public function putAction($id, Request $request)
+    /**
+     * @Security("is_granted('edit', topic)")
+     */
+    public function putAction(Topic $topic, Request $request)
     {
-        if (0 === strpos($request->headers->get('Content-Type'), 'application/json')) {
-            $data = json_decode($request->getContent(), true);
-        }
-        else
-        {
-            $view = $this->view(array(), 301);
-            return $this->handleView($view);
-        }
-        
-        $em = $this->getDoctrine()->getManager();
-        $topic = $em
-            ->getRepository('GSApiBundle:Topic')
-            ->find($id)
-            ;
-        if(! $this->setTopicData($em, $topic, $data))
-        {
-            $view = $this->view(array(), 301);
-            return $this->handleView($view);
-        }
+        $form = $this->get('gsapi.form_generator')->getTopicForm($topic, 'put_topic', 'PUT');
+        $form->handleRequest($request);
 
-        $em->flush();
+        if ($form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            $em->flush();
+
+            $view = $this->view(null, 204);
+            
+        } else {
+            $view = $this->get('gsapi.form_generator')->getFormView($form);
+        }
+        return $this->handleView($view);
     }
     
-    private function setTopicData($em, &$topic, $data)
+    /**
+     * @Security("is_granted('view', topic)")
+     */
+    public function getRegistrationsAction(Topic $topic)
     {
-        $topic->setTitle($data['title']);
-        $topic->setDescription($data['description']);
-        $topic->setState($data['state']);
-
-        // $days = array('Sunday','Monday','Tuesday','Wednesday','Thursday','Friday','Saturday'); 
-        $topic->setDay($data['day']);
-        
-        $topic->setStartTime(new \DateTime($data['startTime']));
-        $topic->setEndTime(new \DateTime($data['endTime']));
-
-        $address = new Address();
-        $topic->setAddress($address);
-        
-        $category = $em
-            ->getRepository('GSApiBundle:Category')
-            ->find($data['categoryId']);
-        $activity = $em
-            ->getRepository('GSApiBundle:Activity')
-            ->find($data['activityId']);
-        if ($activity === null || $category === null) {
-            return false;
-        }
-        $activity->addTopic($topic);
-        $topic->setCategory($category);
-        return true;
-    }
-
-    public function getRegistrationsAction($id)
-    {
-        $topic = $this->getDoctrine()->getManager()
-            ->getRepository('GSApiBundle:Topic')
-            ->find($id)
-            ;
-
         $registrations = $this->getDoctrine()->getManager()
                 ->getRepository('GSApiBundle:Registration')
                 ->findBy(array('topic' => $topic))
                 ;
         
         $view = $this->view($registrations, 200);
+        return $this->handleView($view);
+    }
+    
+    /**
+     * @Security("has_role('ROLE_USER')")
+     */
+    public function newRegistrationAction(Topic $topic)
+    {
+        $registration = new Registration();
+        $registration->setTopic($topic);
+        $this->denyAccessUnlessGranted('create', $registration);
+        $form = $this->get('gsapi.form_generator')->getRegistrationForm($registration, 'post_registration');
+        $view = $this->get('gsapi.form_generator')->getFormView($form);
         return $this->handleView($view);
     }
